http://localhost:8888/notebooks/Machine%20Learning%20Project.ipynb#
My project is based on Stroke data set. Stroke is one of the major disease which can affect human beings. It is also called as brain attack. It occurs when something blocks blood supply to the part of the brain. My objective of this project is to predict the chance of occurrence by using machine learning algorithms. I have used Logistic Regression for my project to find the accuracy.